Simplify shouldEnableStrictMode on newer OSes https://bugs.webkit.org/show_bug.cgi?id=259218 rdar://111926152
Reviewed by Chris Dumez and David Kilzer.

There is an app that is doing some tricky things with the ObjC runtime,
and the containsObject: calls with different classes that exist in the
shouldEnableStrictMode function are causing it to later crash deep inside
the application's own code.  Luckily, we don't actually need to do these
checks on newer OSes where there are only code paths that return true
if strictSecureDecodingForAllObjCEnabled is true.

Manually verified this fixes the app, and it also should marginally
increase performance by removing unnecessary checks.

* Source/WebKit/Shared/Cocoa/ArgumentCodersCocoa.mm:
(IPC::shouldEnableStrictMode):

Canonical link: https://commits.webkit.org/266066@main