Fix SQLiteDatabase::turnOnIncrementalAutoVacuum() https://bugs.webkit.org/show_bug.cgi?id=259284 rdar://112475480
Reviewed by Chris Dumez.

PulkoMandy from the Haiku port found that turnOnIncrementalAutoVacuum is broken again on trunk due
to 258673@main. (It was previously fixed in 241334@main.)

The issue is that we need to call lastError() immediately after the call to columnInt. What we were
doing was calling it immediatley after the destructor for the PRAGMA auto_vacuum statement. So
instead of checking the return code for the statement running, we were checking the return code for
finalizing the statement.

* Source/WebCore/platform/sql/SQLiteDatabase.cpp:
(WebCore::SQLiteDatabase::turnOnIncrementalAutoVacuum):

Canonical link: https://commits.webkit.org/266139@main