Regression(265678@main) Unable to send a MessagePort to an AudioWorklet https://bugs.webkit.org/show_bug.cgi?id=259383 rdar://112624534
Reviewed by Dean Jackson.

In 265678@main, we started validating types being sent via postMessage() to
AudioWorklets, to make sure that these types are actually supposed to be
exposed to AudioWorklet contexts.

Unfortunately, the allow-list was missing the MessagePort type, which is
supposed to be exposed to AudioWorklets. As a result, we were no longer able
to send MessagePorts to AudioWorklets.

This broke the following experiment:
- https://artsandculture.google.com/experiment/viola-the-bird/nAEJVwNkp-FnrQ

* LayoutTests/http/wpt/webaudio/the-audio-api/the-audioworklet-interface/audioworklet-postmessage-message-port.https-expected.txt: Added.
* LayoutTests/http/wpt/webaudio/the-audio-api/the-audioworklet-interface/audioworklet-postmessage-message-port.https.html: Added.
* LayoutTests/http/wpt/webaudio/the-audio-api/the-audioworklet-interface/processors/message-port-processor.js: Added.
* Source/WebCore/bindings/js/SerializedScriptValue.cpp:
(WebCore::isTypeExposedToGlobalObject):

Canonical link: https://commits.webkit.org/266197@main