[ Monterey+ WK2 ] http/wpt/service-workers/cache-mode-hard-reload-serviceworker.html is a flaky TEXT failure https://bugs.webkit.org/show_bug.cgi?id=259382 rdar://112630912
Reviewed by Alex Christensen.

fourth, fifth and sixth script resources were expected to be in the memory cache thus avoiding the service worker since cache mode should be default.
But memory cache may not always kick in in which case the loads will hit service worker.
In that case, it is important to validate that the cache mode is default but we do not log those loads since they might not reliably hit the service worker.

* LayoutTests/http/wpt/service-workers/cache-mode-hard-reload-serviceworker.html:

Canonical link: https://commits.webkit.org/266240@main