[iOS] REGRESSION(265596@main) Captions fail to render in video-fullscreen mode https://bugs.webkit.org/show_bug.cgi?id=259438 rdar://112496682
Reviewed by Eric Carlson.

VTTCues now use `cqw` and `cqh` units to define the height of individual cues. However,
the text-track-container did not establish a container for those units, so the height of
the still-inline <video> element was used to generate the height of cues. In the case of
very small <video> elements, this resulted in no visible cues being painted.

Instead, mark the text-track-container as a basis for inline and block size calculations.

* Source/WebCore/Modules/modern-media-controls/controls/text-tracks.css:
(video::-webkit-media-text-track-container):

Canonical link: https://commits.webkit.org/266243@main